biz_auth/requirements.txt
Django==5.1.1
Django>=4.2
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-decouple
gunicorn
pillow
django-widget-tweaks
drf-yasg
elasticsearch
celery
drf-spectacular
django-jazzmin
django-otp
psycopg[binary]
django-widget-tweaks